+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `@homebridge-plugins/homebridge-meross` will be documented in this file.
 
-## v10.19.3 (Pending Release)
+## v10.20.0 (2026-08-16)
 
 ### Changed